Order posts by the reverse-publishing order
In the implementation of the getTimeline method I have used the spread
operator on the timeline before calling reverse on it. The reason that I
have done this is that the revers method operates IN-PLACE, and thus
mutates the array. Using the spread operator gives us a copy that we can
mutate freely.