Add a wall, containing the posts of followed users
I added an optional parameter to the publishPost method to allow for
posting with a different date because I failed to figure out how to mock
the global date object within the tests. Without some way of changing
the date, all of the posts would have the same time down to the
millisecond, and sorting them by date would have no effect.